+++ MessagePipe.sln
@@ -44,11 +44,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessagePipe.AnalyzerTestApp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessagePipe.Analyzer.Tests", "tests\MessagePipe.Analyzer.Tests\MessagePipe.Analyzer.Tests.csproj", "{64A0C890-301B-49D9-95C8-8CAFB53DCFDC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessagePipe.InProcess", "src\MessagePipe.InProcess\MessagePipe.InProcess.csproj", "{10E626FF-DE92-4784-A9EB-F5A6BE26D1D3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessagePipe.Interprocess", "src\MessagePipe.Interprocess\MessagePipe.Interprocess.csproj", "{10E626FF-DE92-4784-A9EB-F5A6BE26D1D3}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessagePipe.InProcess.Tests", "tests\MessagePipe.InProcess.Tests\MessagePipe.InProcess.Tests.csproj", "{3B2C3908-E9EB-43F3-9340-3CF5EFBA2F09}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessagePipe.Interprocess.Tests", "tests\MessagePipe.Interprocess.Tests\MessagePipe.Interprocess.Tests.csproj", "{3B2C3908-E9EB-43F3-9340-3CF5EFBA2F09}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InProcessServer", "sandbox\InProcessServer\InProcessServer.csproj", "{0E645BF9-3464-4856-A624-FEFCF0050220}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InterprocessServer", "sandbox\InterprocessServer\InterprocessServer.csproj", "{0E645BF9-3464-4856-A624-FEFCF0050220}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
