+++ sandbox/InterprocessServer/InterprocessServer.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
@@ -10,7 +10,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\MessagePipe.InProcess\MessagePipe.InProcess.csproj" />
+    <ProjectReference Include="..\..\src\MessagePipe.Interprocess\MessagePipe.Interprocess.csproj" />
     <ProjectReference Include="..\..\src\MessagePipe\MessagePipe.csproj" />
   </ItemGroup>
