+++ src/MessagePipe.Interprocess/Icon.png
